+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+ - Add support for environment secrets and variables ([#537](https://github.com/eclipse-csi/otterdog/issues/537))
+ - Add support for `prevent_self_review` on environments ([#680](https://github.com/eclipse-csi/otterdog/issues/680))
+
 ## [1.3.4] - 29/06/2026
 
 ### Fixed